+++ integration-testing/test/cl_node/docker_node.py
@@ -77,9 +77,6 @@ class DockerNode(LoggingDockerBase):
 
         self.deploy_dir = tempfile.mkdtemp(dir="/tmp", prefix='deploy_')
         self.create_resources_dir()
-        # if self.config.network is None:
-        #     self.config.network = f'casperlabs_net_{self.config.number}_{random_string(5)}'
-        #     self.config.docker_client.networks.create(self.network, driver='bridge')
 
         commands = self.container_command
         logging.info(f'{self.container_name} commands: {commands}')
@@ -185,7 +182,7 @@ class DockerNode(LoggingDockerBase):
                 network=self.network,
                 volumes=volumes
             ).decode('utf-8')
-            logging.info(f"OUTPUT {self.container_name} {output}")
+            logging.debug(f"OUTPUT {self.container_name} {output}")
             return output
         except ContainerError as err:
             logging.warning(f"EXITED code={err.exit_status} command='{err.command}' stderr='{err.stderr}'")
